# doc/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/cart": {
      "get": {
        "operationId": "CartController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CartItem"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "查看购物车",
        "tags": [
          "购物车"
        ]
      },
      "delete": {
        "operationId": "CartController_clearCart",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "清空购物车",
        "tags": [
          "购物车"
        ]
      }
    },
    "/api/cart/items": {
      "post": {
        "operationId": "CartController_addItem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCartItemDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "添加商品到购物车",
        "tags": [
          "购物车"
        ]
      }
    },
    "/api/cart/items/{productId}": {
      "put": {
        "operationId": "CartController_updateQuantity",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCartItemDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CartItem"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "修改购物车商品数量",
        "tags": [
          "购物车"
        ]
      },
      "delete": {
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "删除购物车中某商品",
        "tags": [
          "购物车"
        ]
      }
    },
    "/api/orders": {
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "创建订单（从购物车结算）",
        "tags": [
          "订单"
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "我的订单列表",
        "tags": [
          "订单"
        ]
      }
    },
    "/api/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "订单详情",
        "tags": [
          "订单"
        ]
      }
    },
    "/api/orders/{id}/pay": {
      "post": {
        "operationId": "OrdersController_pay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "模拟支付",
        "tags": [
          "订单"
        ]
      }
    },
...
    "/api/admin/orders/{id}": {
      "get": {
        "operationId": "AdminOrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "summary": "订单详情",
        "tags": [
          "后台管理 - 订单"
        ]
      }
    },
...
  },
  "info": {
    "title": "电商 API",
    "description": "NestJS 电商系统 REST API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "role": {
            "enum": [
              "customer",
              "admin"
            ],
            "type": "string",
            "description": "角色"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "passwordHash": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role",
          "createdAt",
          "updatedAt",
          "passwordHash"
        ]
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "分类ID"
          },
          "name": {
            "type": "string",
            "description": "分类名称"
          },
          "description": {
            "type": "string",
            "description": "分类描述"
          },
          "parentId": {
            "type": "number",
            "description": "父分类ID"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "parent": {
            "$ref": "#/components/schemas/Category"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Category"
            }
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt",
          "parent",
          "children"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "商品ID"
          },
          "name": {
            "type": "string",
            "description": "商品名称"
          },
          "description": {
            "type": "string",
            "description": "商品描述"
          },
          "price": {
            "type": "number",
            "description": "价格"
          },
          "stock": {
            "type": "number",
            "description": "库存"
          },
          "imageUrl": {
            "type": "string",
            "description": "主图URL"
          },
          "categoryId": {
            "type": "number",
            "description": "所属分类ID"
          },
          "isActive": {
            "type": "boolean",
            "description": "是否上架"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "stock",
          "categoryId",
          "isActive",
          "createdAt",
          "updatedAt",
          "category"
        ]
      },
      "CartItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "购物车项ID"
          },
          "productId": {
            "type": "number",
            "description": "商品ID"
          },
          "quantity": {
            "type": "number",
            "description": "数量"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "userId": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "required": [
          "id",
          "productId",
          "quantity",
          "createdAt",
          "updatedAt",
          "userId",
          "user",
          "product"
        ]
      },
      "AddCartItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "description": "商品ID",
            "example": 1
          },
          "quantity": {
            "type": "number",
            "default": 1,
            "minimum": 1,
            "description": "数量",
            "example": 2
          }
        },
        "required": [
          "productId"
        ]
      },
      "UpdateCartItemDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "number",
            "minimum": 1,
            "description": "数量",
            "example": 3
          }
        },
        "required": [
          "quantity"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "shippingAddress": {
            "type": "string",
            "maxLength": 500,
            "description": "收货地址",
            "example": "北京市朝阳区XX路100号"
          },
          "remark": {
            "type": "string",
            "maxLength": 500,
            "description": "备注",
            "example": "请尽快发货"
          }
        },
        "required": [
          "shippingAddress"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "订单ID"
          },
          "orderNo": {
            "type": "string",
            "description": "订单号"
          },
          "totalAmount": {
            "type": "number",
            "description": "订单总金额"
          },
          "status": {
            "enum": [
              "pending",
              "confirmed",
              "shipped",
              "completed",
              "cancelled"
            ],
            "type": "string",
            "description": "订单状态"
          },
          "paymentStatus": {
            "enum": [
              "unpaid",
              "paid"
            ],
            "type": "string",
            "description": "支付状态"
          },
          "shippingAddress": {
            "type": "string",
            "description": "收货地址"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "paidAt": {
            "format": "date-time",
            "type": "string",
            "description": "支付时间"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新时间"
          },
          "userId": {
            "type": "number"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          }
        },
        "required": [
          "id",
          "orderNo",
          "totalAmount",
          "status",
          "paymentStatus",
          "shippingAddress",
          "createdAt",
          "updatedAt",
          "userId",
          "user",
          "items"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "订单明细ID"
          },
          "productName": {
            "type": "string",
            "description": "商品名称（快照）"
          },
          "unitPrice": {
            "type": "number",
            "description": "单价（快照）"
          },
          "quantity": {
            "type": "number",
            "description": "数量"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "创建时间"
          },
          "orderId": {
            "type": "number"
          },
          "order": {
            "$ref": "#/components/schemas/Order"
          },
          "productId": {
            "type": "number"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          }
        },
        "required": [
          "id",
          "productName",
          "unitPrice",
          "quantity",
          "createdAt",
          "orderId",
          "order",
          "productId",
          "product"
        ]
      },
...
    }
  }
}